requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
Click==7.0
coverage==4.5.2
Django==2.1.4
django-configurations==2.1
django-cors-headers==2.4.0
django-extensions==2.1.4
django-indexer==0.3.0
django-paging==0.2.5
django-rq==1.2.0
django-templatetag-sugar==1.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
dnspython==1.16.0
docutils==0.14
eventlet==0.24.1
greenlet==0.4.15
lockfile==0.12.2
monotonic==1.5
more-itertools==5.0.0
pluggy==0.8.0
py==1.7.0
PyJWT==1.7.1
pytest==4.1.0
pytest-cov==2.6.1
pytest-django==3.4.5
python-daemon==2.2.0
pytz==2018.7
raven==6.10.0
redis==3.0.1
rq==0.12.0
sentry-sdk==0.6.6
six==1.11.0
South==1.0.2
urllib3==1.24.1
